backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
anthropic>=0.18.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
websockets>=12.0
